requirements.txt
kafka-python==2.0.2
flask==2.0.1
pandas==1.3.3
plotly==5.3.1